requirements.txt
openai>=1.40.0
pydantic>=2.6.0
python-dotenv>=1.0.0
pillow>=10.0.0
requests>=2.31.0
streamlit>=1.32.0
rich>=13.7.0
pytest>=8.0.0
